Check if pagurus if found before starting

+++ test/integration/shifter_exec.sh
@@ -58,18 +58,24 @@ else
     ID=$1
 fi
 
-# Start monitoring running, get monitor pid and sleep
-pagurus -u $USER -o $PWD/stats.csv &
-export PID=$!
-sleep 2
+# Check if monitoring script is installed
+if [ -x "$(command -v pagurus)" ]; then
+    # Start monitoring running, get monitor pid and sleep
+    pagurus -u $USER -o $PWD/stats.csv &
+    PID=$!
+    sleep 2
+fi
 
 # Run container script and catch exit code
 shifter --image=$ID -V $2:$3 $4 $5
 export EXIT_CODE=$?
 
-# Kill pagurus monitoring and wait for it to write stats.csv
-kill $PID
-sleep 5
+# If PID is set then kill it
+if [ -n "$PID" ]; then
+    # Kill pagurus monitoring and wait for it to write stats.csv
+    kill $PID
+    sleep 2
+fi
 
 # Return with container exit code
 exit $EXIT_CODE
